Added component catalog page for react
