+++ docs/src/markdownDocs/components.mdx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { ComponentCatalogGrids } from '../pages/ComponentCatalogGrids';
+
+{/* KEYWORDS: components typography icon storybook devdoc catalogue catalog */}
+
+# Brightlayer UI Components
+
+Brightlayer UI offers custom component libraries to simplify building your application by implementing common use cases in Brightlayer UI and eliminating the need for multiple teams to build their own components. These components are supplementary to basic components offered by [MUI](https://mui.com/material-ui/getting-started/supported-components/).
+
+Brightlayer UI react components are available as NPM package:
+
+-   [@brightlayer-ui/react-components](https://www.npmjs.com/package/@brightlayer-ui/react-components)
+
+<ComponentCatalogGrids />
