+++ docs/src/markdownDocs/index.tsx
@@ -1,5 +1,7 @@
 import React from './react.mdx';
 import Environment from './environment.mdx';
+import Components from './components.mdx';
 
 export { React };
 export { Environment };
+export { Components };
