--- docs/src/pages/ComponentCatalog.tsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import { ComponentCatalogGrids } from './ComponentCatalog/ComponentCatalogGrids';
-
-export const ComponentCatalog = (): JSX.Element => <ComponentCatalogGrids />;
