--- docs/src/pages/ComponentCatalog/index.tsx
@@ -1,1 +1,0 @@
-export * from './ComponentCatalogLinks';
