+++ docs/src/pages/index.tsx
@@ -1,3 +1,3 @@
 export * from './componentPreviewPage';
 export * from './MarkdownPage';
-export * from './ComponentCatalog';
+export * from './ComponentCatalogGrids';
